refactor(routes): reorganize imports and update paths for consistency
