+++ apps/demo/src/app/components/FeaturesComparison.tsx
