+++ apps/demo/src/app/pages/color-palette/ColorPalette.tsx
